+++ service/client/artists/queries.ts
